Minor reorganization
Move AI Cores into new AICores folder, and AI Movement Controllers into Movement (renamed from MovementAI)
