+++ TAC_AI/AI/Movement/AICores/AircraftUtils.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using System.Reflection;
 using UnityEngine;
 
-namespace TAC_AI.AI.MovementAI
+namespace TAC_AI.AI.Movement.AICores
 {
     internal class AircraftUtils
     {
