+++ TAC_AI/AI/Movement/IMovementAIController.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using TAC_AI.AI.MovementAI;
+using TAC_AI.AI.Movement.AICores;
+
 
 namespace TAC_AI.AI {
