+++ TAC_AI/Enemy/RPathfinding.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using UnityEngine;
+using TAC_AI.AI.Movement;
 
 namespace TAC_AI.AI.Enemy
 {
